--- core/src/main/kotlin/dsl/Status.kt
@@ -1,30 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2019-2020 Pierre Leresteux.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dsl
-
-/**
- * Represent the status of the changeset.
- */
-enum class Status {
-    // execution in progress
-    EXECUTE,
-    // Execution is done with a correct status
-    OK,
-    // Execution is done but it has failed
-    KO
-}
