+++ core/src/main/kotlin/io/saagie/updatarium/dsl/Changelog.kt
@@ -15,25 +15,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import io.saagie.updatarium.dsl.action.BashScriptAction
-import io.saagie.updatarium.dsl.changeSet
-import io.saagie.updatarium.dsl.changelog
+package io.saagie.updatarium.dsl
 
-changelog {
-    changesets {
-        +changeSet {
-            id = "ChangeSet-bash-1"
-            author = "Bash"
-            actions {
-                +BashScriptAction(
-                    script = """
-curl -I https://httpbin.org/get | grep -i Server &&\
-pwd &&\
-export | grep " PWD"
-""".trimIndent(),
-                    workingDir = "/tmp"
-                )
-            }
+import com.autodsl.annotation.AutoDsl
+import io.saagie.updatarium.persist.DefaultPersistEngine
+import io.saagie.updatarium.persist.PersistEngine
+
+@AutoDsl
+data class Changelog(var changesets: List<ChangeSet> = mutableListOf()) {
+
+    /**
+     * It will execute each changesets present in this changelog sequentially.
+     */
+    fun execute(engine: PersistEngine) {
+        engine.checkConnection()
+        changesets.forEach {
+            it.execute(engine)
         }
     }
 }
+
