+++ engine-bash/src/main/kotlin/io/saagie/updatarium/dsl/action/BashScriptAction.kt
@@ -15,25 +15,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import io.saagie.updatarium.dsl.action.BashScriptAction
-import io.saagie.updatarium.dsl.changeSet
-import io.saagie.updatarium.dsl.changelog
+package io.saagie.updatarium.dsl.action
 
-changelog {
-    changesets {
-        +changeSet {
-            id = "ChangeSet-bash-1"
-            author = "Bash"
-            actions {
-                +BashScriptAction(
-                    script = """
-curl -I https://httpbin.org/get | grep -i Server &&\
-pwd &&\
-export | grep " PWD"
-""".trimIndent(),
-                    workingDir = "/tmp"
-                )
-            }
-        }
+import com.autodsl.annotation.AutoDsl
+import io.saagie.updatarium.engine.bash.BashEngine
+import java.util.concurrent.TimeUnit
+
+@AutoDsl
+data class BashScriptAction(
+    val script: String,
+    val workingDir: String = ".",
+    val timeoutAmount: Long = 60,
+    val timeoutUnit: TimeUnit = TimeUnit.SECONDS
+) : Action() {
+    val bashEngine = BashEngine()
+
+    override fun execute() {
+        bashEngine.runCommand(this)
     }
 }
