--- engine-httpclient/src/main/kotlin/dsl/action/HttpScriptAction.kt
@@ -1,31 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2019-2020 Pierre Leresteux.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dsl.action
-
-import com.autodsl.annotation.AutoDsl
-import com.github.kittinunf.fuel.core.FuelManager
-
-@AutoDsl
-class HttpScriptAction(val f: (httpScriptAction: HttpScriptAction) -> Unit) : Action() {
-
-    val restClient = FuelManager.instance
-
-    override fun execute() {
-        f(this)
-    }
-}
