+++ engine-httpclient/src/main/kotlin/io/saagie/updatarium/dsl/action/HttpScriptAction.kt
@@ -15,25 +15,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import io.saagie.updatarium.dsl.action.BashScriptAction
-import io.saagie.updatarium.dsl.changeSet
-import io.saagie.updatarium.dsl.changelog
+package io.saagie.updatarium.dsl.action
 
-changelog {
-    changesets {
-        +changeSet {
-            id = "ChangeSet-bash-1"
-            author = "Bash"
-            actions {
-                +BashScriptAction(
-                    script = """
-curl -I https://httpbin.org/get | grep -i Server &&\
-pwd &&\
-export | grep " PWD"
-""".trimIndent(),
-                    workingDir = "/tmp"
-                )
-            }
-        }
+import com.autodsl.annotation.AutoDsl
+import com.github.kittinunf.fuel.core.FuelManager
+
+@AutoDsl
+class HttpScriptAction(val f: (httpScriptAction: HttpScriptAction) -> Unit) : Action() {
+
+    val restClient = FuelManager.instance
+
+    override fun execute() {
+        f(this)
     }
 }
